qcacld-3.0: Enable monitor mode support
If the device's wifi/mobile data is shared via hotspot with another device,
and if the host device enters suspend, the modem crashes and wifi/hotspot
turns off. A thorough checkup in qcacld's defconfig pointed out that this
is due to disabling monitor mode. Enable it.

NOTE: This was found on SM7325 powered POCO X5 Pro. To avoid potential issue
in the future, enable on our current platform as well.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: Santhosh <santhosh.user.why.red@gmail.com>